pyproject.toml
[project]
name = "ingestion_pipeline_simple"
requires-python = ">=3.10,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.12.4",
    "markitdown[all]>=0.1.4",
    "pypandoc>=1.16.2",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "ingestion_pipeline_simple"
registry_modules = [
    "ingestion_pipeline_simple.components.*",
]


